Create RubiksCube.cpp
This is my first commit that I made in this repository.
This is the broad strokes of what I want it to do.
Some things that I would like to work on is that all of these moves are basically the same when you break it down. So there should be a way that is much simpler than this solution. 

My main goal is to find a way that I only have
void move(int FACE);
and when calling
move(UP); ECT. 
This would make things a lot simpler.