commands changed
I changed the commands so they seam to work the way I want 
You can type in an algorithm, and it will preform all of the move example:
"B' U2 R" would rotate the Bottom Face Counter Clockwise, Up Face twice, and Right Clockwise  
I will be able to use this when I am trying to solve the cube. 

My next goal is to make sure that all of my moves are moving the way I want them to. 
I am pretty sure that I miss typed in some of the variables in the ATVM (Auto Variable Move)